Done the following menu entries: 2 5 6
The item 4 works as long as we set total with the right amount.
the item 3 is about integrating the code with perl
the item 1 works around the directory files (UNIX ONLY)

we need 1 to count the emails so that item 4 is doable. Once we can
remove and count files the last thing will be integrate with perl.